moved proptypes
