+++ src/components/m-datepicker/index.js
@@ -1,5 +1,8 @@
+import PropTypes from 'prop-types';
+
 import BaseComponentGlobal from '../../js/abstract/base-component-global';
 import defineOnce from '../../js/define-once';
+import { TODAY } from '../../js/date';
 // import the styles used for this component
 import styles from './index.scss';
 // import the template used for this component
@@ -9,8 +12,26 @@ import Datepicker from './js/datepicker';
 
 import getAttribute from '../../js/get-attribute';
 
+const startType = PropTypes.oneOfType([
+  PropTypes.number,
+  PropTypes.oneOf([TODAY]),
+]);
+
 class AXADatepicker extends BaseComponentGlobal {
   static tagName = 'axa-m-datepicker'
+  static propTypes = {
+    classes: PropTypes.string,
+    buttonOk: PropTypes.string,
+    buttonCancel: PropTypes.string,
+    locale: PropTypes.string,
+    value: PropTypes.string,
+    startYear: startType,
+    startMonth: startType,
+    selectedDay: PropTypes.number,
+    lowerEndYear: PropTypes.number,
+    higherEndYear: PropTypes.number,
+    outputIso: PropTypes.string,
+  }
 
   // Specify observed attributes so that attributeChangedCallback will work,
   // this is essential for external re-rendering trigger.
