+++ src/components/m-dropdown/index.js
@@ -5,8 +5,8 @@ import styles from './index.scss';
 import template from './_template';
 import BaseComponentGlobal from '../../js/abstract/base-component-global';
 import defineOnce from '../../js/define-once';
-import urlPropType from '../../js/url-prop-type';
-import valuePropType from '../../js/value-prop-type';
+import urlPropType from '../../js/prop-types/url-prop-type';
+import valuePropType from '../../js/prop-types/value-prop-type';
 import DropDown from './js/drop-down';
