+++ src/components/m-header-mobile-others/index.js
@@ -4,7 +4,7 @@ import styles from './index.scss';
 import template from './_template';
 import BaseComponentGlobal from '../../js/abstract/base-component-global';
 import defineOnce from '../../js/define-once';
-import urlPropType from '../../js/url-prop-type';
+import urlPropType from '../../js/prop-types/url-prop-type';
 
 class AXAHeaderMobileOthers extends BaseComponentGlobal {
   static tagName = 'axa-header-mobile-others'
