+++ src/components/m-header-navigation/index.js
@@ -7,7 +7,7 @@ import Stroke from './js/stroke';
 import HeaderNavigation from './js/header-navigation';
 import BaseComponentGlobal from '../../js/abstract/base-component-global';
 import defineOnce from '../../js/define-once';
-import urlPropType from '../../js/url-prop-type';
+import urlPropType from '../../js/prop-types/url-prop-type';
 
 class AXAHeaderNavigation extends BaseComponentGlobal {
   static tagName = 'axa-header-navigation'
