+++ src/components/m-header-sub-navigation/index.js
@@ -5,7 +5,7 @@ import styles from './index.scss';
 import template from './_template';
 import BaseComponentGlobal from '../../js/abstract/base-component-global';
 import defineOnce from '../../js/define-once';
-import urlPropType from '../../js/url-prop-type';
+import urlPropType from '../../js/prop-types/url-prop-type';
 
 class AXAHeaderSubNavigation extends BaseComponentGlobal {
   static tagName = 'axa-header-sub-navigation'
