+++ src/components/o-commercial-hero-cover/index.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 
 import BaseComponentGlobal from '../../js/abstract/base-component-global';
 import defineOnce from '../../js/define-once';
-import urlPropType from '../../js/url-prop-type';
+import urlPropType from '../../js/prop-types/url-prop-type';
 import styles from './index.scss';
 import template from './_template';
