+++ src/js/prop-types/value-prop-type.js
